Radarr Monitor existing movies
This is a small change that allows you to re-monitor a movie in Radarr that was previously unmonitored.

The use case here is:
You have something in radarr and radarr is set to unmonitor files that have been deleted. A file gets deleted but it reappears on one of the popular lists. In this circumstance I want to monitor it again and get it next time it appears in the RSS.